Create update-interface-configs-from-current-status.py
This script should be ran in a cron job every 1-5 minutes. If you have a workstation that has multiple nics, and you plug the ethernet cable into a nic that has the 'ONBOOT' set to no, then this script will detect a difference between the current status of that link, and the persistent config on disk. It will update the config to turn on boot, and restart the network service.

I imagine NetworkManager would handle this, but NetworkManager is turned off because it causes other problems.